+++ javacore-song-task/src/main/java/META-INF/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: ru.liga.App
