+++ javacore-song-task/src/main/resources/Logback.xml
@@ -21,10 +21,10 @@
             <maxIndex>10</maxIndex>
         </rollingPolicy>
         <triggeringPolicy class="ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
-            <maxFileSize>10KB</maxFileSize>
+            <maxFileSize>100KB</maxFileSize>
         </triggeringPolicy>
         <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%-5level] - %msg%n</pattern>
+            <pattern>%msg%n</pattern>
         </encoder>
     </appender>
